fixed mappers

+++ src/main/java/com/fontana/backend/devicesType/light/dto/LightDTO.java
@@ -6,14 +6,14 @@ import lombok.Data;
 public class LightDTO {
 
     private String name;
-    private int colorR;
-    private int colorG;
-    private int colorB;
+    private byte colorR;
+    private byte colorG;
+    private byte colorB;
 
     public LightDTO() {
     }
 
-    public LightDTO(String name, int colorR, int colorG, int colorB) {
+    public LightDTO(String name, byte colorR, byte colorG, byte colorB) {
         this.name = name;
         this.colorR = colorR;
         this.colorG = colorG;
